prevent sqli

+++ message.php
@@ -12,7 +12,7 @@ if (!$conn) {
 }
 mysqli_select_db($conn,$MYSQL_DB); 
 if (isset($_GET['id'])) {
-    $id = mysqli_real_escape_string($conn,$_GET['id']);
+    $id = (int) mysqli_real_escape_string($conn,$_GET['id']);
     $sql = "SELECT * FROM UserInformation WHERE id = $id";
     $result = $conn -> query($sql);
 }
